backend/ai_core/pyproject.toml
[project]
name = "yc-hackathon-ai-core"
version = "0.1.0"
description = "AI core logic"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.66.0",
    "pydantic>=2.12.4",
    "yc-hackathon-shared",
]

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"

[tool.uv.sources]
yc-hackathon-shared = { workspace = true }
